# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Movie API
  version: 1.0.0
servers:
- url: https://localhost:8080/api/v1
...
paths:
...
  /showtimes:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Showtime'
                type: array
          description: OK
      summary: List all showtimes
      tags:
      - Showtimes
      x-accepts: application/json
      x-tags:
      - tag: Showtimes
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Showtime'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Showtime'
          description: Created
      summary: Create a new showtime
      tags:
      - Showtimes
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Showtimes
  /showtimes/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_showtimes__id__get_404_response'
          description: The specified showtime was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_movies__id__put_500_response'
          description: An internal server error occurred
      summary: Delete a specific showtime
      tags:
      - Showtimes
      x-accepts: application/json
      x-tags:
      - tag: Showtimes
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Showtime'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_showtimes__id__get_404_response'
          description: The specified showtime was not found
      summary: Get details of a specific showtime
      tags:
      - Showtimes
      x-accepts: application/json
      x-tags:
      - tag: Showtimes
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Showtime'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Showtime'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_showtimes__id__get_404_response'
          description: The specified showtime was not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_showtimes__id__put_409_response'
          description: A showtime with the same ID already exists
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_movies__id__put_500_response'
          description: An internal server error occurred
      summary: Update details of a specific showtime
      tags:
      - Showtimes
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Showtimes
...
components:
...
  schemas:
...
    Showtime:
      example:
        theaterId: 1
        price: 5.962133916683182
        showtimeId: 0
        movieId: 6
        startTime: 2000-01-23T04:56:07.000+00:00
        endTime: 2000-01-23T04:56:07.000+00:00
      properties:
        showtimeId:
          description: Unique identifier for the showtime
          format: int64
          type: integer
        movieId:
          description: ID of the movie being shown
          format: int64
          type: integer
        theaterId:
          description: ID of the theater where the showtime is taking place
          format: int64
          type: integer
        startTime:
          description: Start time of the showtime
          format: date-time
          type: string
        endTime:
          description: End time of the showtime
          format: date-time
          type: string
        price:
          description: Ticket price for the showtime
          format: double
          type: number
      type: object
...
    _movies__id__put_500_response:
      properties:
        message:
          description: Generic error message
          type: string
      type: object
    _showtimes__id__get_404_response:
      properties:
        message:
          description: Error message indicating showtime not found
          type: string
      type: object
    _showtimes__id__put_409_response:
      properties:
        message:
          description: Error message indicating duplicate showtime ID
          type: string
      type: object
...